requirement.txt
langchain
langchain-openai
python-dotenv
fastapi
uvicorn
requests